Refactor hypothesis strategies
